% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumar.R
\name{sumar}
\alias{sumar}
\title{Sum two elements}
\usage{
sumar(x, y, negative = FALSE)
}
\arguments{
\item{x}{First element to sum.}

\item{y}{Second element to sum.}

\item{negative}{TRUE if the elements to subtract.}
}
\value{
The sum of the two elements.
}
\description{
Sum two elements
}
\examples{
x <- 3
y <- 3
sumar(x,y)
}
